Add profile page step
