+++ xcode/Subconscious/Subconscious.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -6,7 +6,7 @@
       "location" : "https://github.com/subconsciousnetwork/noosphere",
       "state" : {
         "branch" : "main",
-        "revision" : "609a72d23ba23d1d9541693a276ccfcc0d983215"
+        "revision" : "b404ec0d117e2467bfbe4a3bda4253e1c57f584e"
       }
     },
     {
